refactor: PlainDate - phase 1 - migrate util methods
