+++ assembly/plaindate.ts
@@ -7,7 +7,7 @@ import { PlainMonthDay } from "./plainmonthday";
 import { PlainTime } from "./plaintime";
 import { PlainYearMonth } from "./plainyearmonth";
 import {
-  addDate,
+  clamp,
   dayOfWeek,
   leapYear,
   dayOfYear,
@@ -15,12 +15,13 @@ import {
   daysInMonth,
   daysInYear,
   toPaddedString,
-  rejectDate,
   checkDateTimeRange,
-  compareTemporalDate,
-  differenceDate,
   coalesce,
   parseISOString,
+  YMD,
+  checkRange,
+  compare,
+  floorDiv,
 } from "./utils";
 
 export class DateLike {
@@ -45,6 +46,47 @@ export class PlainDate {
       }
       throw new TypeError("invalid date type");
     }
+  }
+
+  static balanced(year: i32, month: i32, day: i32): PlainDate {
+    const yearMonth = PlainYearMonth.balanced(year, month);
+
+    year  = yearMonth.year;
+    month = yearMonth.month;
+
+    let daysPerYear = 0;
+    let testYear = month > 2 ? year : year - 1;
+
+    while (((daysPerYear = daysInYear(testYear)), day < -daysPerYear)) {
+      year -= 1;
+      testYear -= 1;
+      day += daysPerYear;
+    }
+
+    testYear += 1;
+
+    while (((daysPerYear = daysInYear(testYear)), day > daysPerYear)) {
+      year += 1;
+      testYear += 1;
+      day -= daysPerYear;
+    }
+
+    while (day < 1) {
+      const yearMonth = PlainYearMonth.balanced(year, month - 1);
+      year  = yearMonth.year;
+      month = yearMonth.month;
+      day  += daysInMonth(year, month);
+    }
+
+    let monthDays = 0;
+    while (monthDays = daysInMonth(year, month), day > monthDays) {
+      const yearMonth = PlainYearMonth.balanced(year, month + 1);
+      year  = yearMonth.year;
+      month = yearMonth.month;
+      day  -= monthDays;
+    }
+
+    return new PlainDate(year, month, day);
   }
 
   @inline
@@ -267,6 +309,220 @@ export class PlainDate {
 
   static compare(a: PlainDate, b: PlainDate): i32 {
     if (a === b) return 0;
-    return compareTemporalDate(a.year, a.month, a.day, b.year, b.month, b.day);
+    return compare(
+      [a.year, a.month, a.day],
+      [b.year, b.month, b.day]
+    );
+  }
+}
+
+// https://github.com/tc39/proposal-temporal/blob/49629f785eee61e9f6641452e01e995f846da3a1/polyfill/lib/ecmascript.mjs#L2984
+function addDate(
+  year: i32,
+  month: i32,
+  day: i32,
+  years: i32,
+  months: i32,
+  weeks: i32,
+  days: i32,
+  overflow: Overflow
+): YMD {
+  year  += years;
+  month += months;
+
+  const yearMonth = PlainYearMonth.balanced(year, month);
+  year  = yearMonth.year;
+  month = yearMonth.month;
+
+  const regulatedDate = regulateDate(year, month, day, overflow);
+  year  = regulatedDate.year;
+  month = regulatedDate.month;
+  day   = regulatedDate.day;
+  day  += days + weeks * 7;
+
+  const balancedDate = PlainDate.balanced(year, month, day);
+  year  = balancedDate.year;
+  month = balancedDate.month;
+  day   = balancedDate.day;
+
+  return { year, month, day };
+}
+
+function rejectDate(year: i32, month: i32, day: i32): void {
+  if (!checkRange(month, 1, 12)) {
+    throw new RangeError("month out of range");
+  }
+  if (!checkRange(day, 1, daysInMonth(year, month))) {
+    throw new RangeError("day out of range");
+  }
+}
+
+// https://github.com/tc39/proposal-temporal/blob/49629f785eee61e9f6641452e01e995f846da3a1/polyfill/lib/ecmascript.mjs#L2617
+function constrainDate(year: i32, month: i32, day: i32): YMD {
+  month = clamp(month, 1, 12);
+  day   = clamp(day, 1, daysInMonth(year, month));
+  return { year, month, day };
+}
+
+// https://github.com/tc39/proposal-temporal/blob/49629f785eee61e9f6641452e01e995f846da3a1/polyfill/lib/ecmascript.mjs#L2617
+function regulateDate(
+  year: i32,
+  month: i32,
+  day: i32,
+  overflow: Overflow
+): YMD {
+  switch (overflow) {
+    case Overflow.Reject:
+      rejectDate(year, month, day);
+      break;
+
+    case Overflow.Constrain:
+      const date = constrainDate(year, month, day);
+      year  = date.year;
+      month = date.month;
+      day   = date.day;
+      break;
+  }
+
+  return { year, month, day };
+}
+
+
+function differenceDate(
+  yr1: i32, mo1: i32, d1: i32,
+  yr2: i32, mo2: i32, d2: i32,
+  largestUnit: TimeComponent = TimeComponent.Days
+): Duration {
+  switch (largestUnit) {
+    case TimeComponent.Years:
+    case TimeComponent.Months: {
+      let sign = -compare(
+        [yr1, mo1, d1],
+        [yr2, mo2, d2]
+      );
+      if (sign == 0) return new Duration();
+
+      let startYear  = yr1;
+      let startMonth = mo1;
+
+      let endYear  = yr2;
+      let endMonth = mo2;
+      let endDay   = d2;
+
+      let years = endYear - startYear;
+      let mid = new PlainDate(yr1, mo1, d1)
+        .add(new Duration(years), Overflow.Constrain);
+      let midSign = -compare(
+        [mid.year, mid.month, mid.day],
+        [yr2, mo2, d2]
+      );
+
+      if (midSign === 0) {
+        return largestUnit === TimeComponent.Years
+          ? new Duration(years)
+          : new Duration(0, years * 12);
+      }
+
+      let months = endMonth - startMonth;
+
+      if (midSign !== sign) {
+        years  -= sign;
+        months += sign * 12;
+      }
+
+      mid = new PlainDate(yr1, mo1, d1)
+        .add(new Duration(years, months), Overflow.Constrain);
+      midSign = -compare(
+        [mid.year, mid.month, mid.day], 
+        [yr2, mo2, d2]
+      );
+
+      if (midSign === 0) {
+        return largestUnit === TimeComponent.Years
+          ? new Duration(years, months)
+          : new Duration(0, months + years * 12);
+      }
+
+      if (midSign !== sign) {
+        // The end date is later in the month than mid date (or earlier for
+        // negative durations). Back up one month.
+        months -= sign;
+
+        if (months === -sign) {
+          years -= sign;
+          months = sign * 11;
+        }
+
+        mid = new PlainDate(yr1, mo1, d1)
+          .add(new Duration(years, months), Overflow.Constrain);
+      }
+
+      let days = endDay - mid.day; // If we get here, months and years are correct (no overflow), and `mid`
+      // is within the range from `start` to `end`. To count the days between
+      // `mid` and `end`, there are 3 cases:
+      // 1) same month: use simple subtraction
+      // 2) end is previous month from intermediate (negative duration)
+      // 3) end is next month from intermediate (positive duration)
+
+      if (mid.month === endMonth && mid.year === endYear) {
+        // 1) same month: use simple subtraction
+      } else if (sign < 0) {
+        // 2) end is previous month from intermediate (negative duration)
+        // Example: intermediate: Feb 1, end: Jan 30, DaysInMonth = 31, days = -2
+        days -= daysInMonth(endYear, endMonth);
+      } else {
+        // 3) end is next month from intermediate (positive duration)
+        // Example: intermediate: Jan 29, end: Feb 1, DaysInMonth = 31, days = 3
+        days += daysInMonth(mid.year, mid.month);
+      }
+
+      if (largestUnit === TimeComponent.Months) {
+        months += years * 12;
+        years = 0;
+      }
+
+      return new Duration(years, months, 0, days);
+    }
+
+    case TimeComponent.Weeks:
+    case TimeComponent.Days: {
+      let neg = compare(
+        [yr1, mo1, d1],
+        [yr2, mo2, d2]
+      ) < 0;
+
+      let smallerYear  = neg ? yr1 : yr2;
+      let smallerMonth = neg ? mo1 : mo2;
+      let smallerDay   = neg ? d1 : d2;
+
+      let largerYear  = neg ? yr2 : yr1;
+      let largerMonth = neg ? mo2 : mo1;
+      let largerDay   = neg ? d2 : d1;
+
+      let sign = neg ? 1 : -1;
+
+      let years = largerYear - smallerYear;
+
+      let days = (
+        dayOfYear(largerYear,  largerMonth,  largerDay) -
+        dayOfYear(smallerYear, smallerMonth, smallerDay)
+      );
+
+      while (years > 0) {
+        days  += daysInYear(smallerYear + years - 1);
+        years -= 1;
+      }
+
+      let weeks = 0;
+      if (largestUnit === TimeComponent.Weeks) {
+        weeks = floorDiv(days, 7);
+        days -= weeks * 7;
+      }
+
+      return new Duration(0, 0, weeks * sign, days * sign);
+    }
+
+    default:
+      throw new Error('differenceDate - cannot support TimeComponent < Days');
   }
 }
