+++ assembly/refactor.md
@@ -1,0 +1,8 @@
+Steps
+
+ - [ ] Move fns outside of utils
+  - [x] Duration
+  - [x] PlainDate
+  - [x] Instant
+ - [ ] Create a seperate utils and dateTimeUtils classes
+ - [ ] refactors fns that have been moved outside of utils
